Set 'active collection' with location hash param
The List view now watches the 'c' param and uses it to determine the
active collection--i.e., the one that is displayed.

Sorting works

See 0cd6552b32ad9ee81a4f3edf0db875b671a2879e for guidelines on
refactoring this change, since other parts of the app will be interested
in the 'active' collection.