Implemented text and bounds filtering
The Proposals collection determines if the query is narrower than the
last query it executed on the server, in which case it calculates the
results (in a fairly expensive, but still cheaper than
hitting-the-network way...) from the local cache. If it is broader than
the last query (true when there was no last query, i.e., on
initialization), then it generates an appropriate query and fetches
results from the server.

Current behavior of the bounds filter may be a bit weird, but it is
serviceable.