Working version of Docker Compose workflow
This replaces the fragile setup that the project previously used, which
relied on a monolithic Docker image that put the Django app and all of
its dependencies in the same container. Data would not persist beyond
the lifetime of the container (by design), meaning that every restart
involved a somewhat lengthy re-migration.  ... Among other disadvantages.

Remaining to be done:

- Import parcels: rather than prepopulate the database with parcels,
  write importer scripts after the fashion of the proposal importers. I
  was hoping that I could write a generic parcel importer, but the
  inputs are likely to be too heterogeneous.

- Run initialization scripts on first container run. They can be
  responsible for populating the database with some parcels (and some
  proposals?)

- Edit /start.sh to use docker-compose.

- Figure out how to set up Celery. Should it run in the same container
  as the web app?