Start script: make sure Postgres is online first
The first time a user runs docker-compose up, Postgres has to do a lot
of work initializing the database, so it remains unavailable for a while
after its container has started. With this change, start.sh checks to
make sure that Postgres is available before attempting to run any
migrations or start Django.