Removed --detach option in celery startup
For reasons I haven't been able to determine, celery does not work
correctly when daemonized, even when all of the configuration is passed
in explicitly. For now, when starting celery in dev mode, just use
bash's built-in forking. Actually, it's nice--stdout and stderr print to
the docker-compose log this way.