Template build script; optional module defaults
script/build_templates.py will be the first stage in the JS build
process. The script assembles all the templates in client/templates and
package them into a "templates" module in client/scripts/build.

Also part of this commit, utils.templateWithUrl first checks the
template module before placing a request.

When a module is included as "optional!<moduleName>" and <moduleName> is
not available, look in config.moduleDefaults.