Support Proposal queries against time range
It's not obvious how this should work. For the time being, it queries by the
time of Proposal creation date. The question is whether that composes
intuitively with the other possible queries a user could make. For example, if
the query is for completed (approved/denied) proposals, should it find proposals
that were completed within that range? If the query is for pending
proposals (i.e., under review), should it find proposals that were still active
during that time range?

Probably "yes" to both, right?

Well, I've convinced myself. But I'm going to implement that in a subsequent
commit, so goodbye for now, O Best Beloved.