Account deactivation by replying to email
I need to follow up on this to see if it's a secure method of processing user
commands in message bodies. I suspect it may not be, since the From header can
be spoofed; but SendGrid may reject spoofed messages out of hand. An alternative
approach, not especially difficult to implement, would be to create unique
Reply-To address for each message that includes a verification code.