Methods for calculating Subscription similarity
Subscription similarity is a float between 0 and 1. If the Subscriptions being
compared have geographic shapes, the value reflects the percentage overlap.

This is not used yet, but the intention is to do the 'right thing' if a user
creates two highly similar Subscriptions; where 'right thing' is to be defined.